+++ dflib-jdbc/src/main/java/org/dflib/jdbc/connector/SqlLoader.java
@@ -25,46 +25,65 @@ public class SqlLoader {
     protected final JdbcConnector connector;
     private final String sql;
 
-    protected int maxRows;
+    protected int limit;
     private int rowSampleSize;
     private Random rowsSampleRandom;
 
     public SqlLoader(JdbcConnector connector, String sql) {
         this.connector = connector;
-        this.maxRows = Integer.MAX_VALUE;
+        this.limit = Integer.MAX_VALUE;
         this.sql = sql;
     }
 
     protected SqlLoader copy() {
         SqlLoader copy = new SqlLoader(connector, sql);
-        copy.maxRows = this.maxRows;
+        copy.limit = this.limit;
         copy.rowSampleSize = this.rowSampleSize;
         copy.rowsSampleRandom = this.rowsSampleRandom;
         return copy;
     }
 
-    public SqlLoader maxRows(int maxRows) {
-        if (this.maxRows == maxRows) {
+    /**
+     * @since 1.0.0-M20
+     */
+    public SqlLoader limit(int maxRows) {
+        if (this.limit == maxRows) {
             return this;
         }
 
         SqlLoader copy = copy();
-        copy.maxRows = maxRows;
+        copy.limit = maxRows;
         return copy;
+    }
+
+    /**
+     * @deprecated in favor of {@link #limit(int)}
+     */
+    @Deprecated(since = "1.0.0-M20", forRemoval = true)
+    public SqlLoader maxRows(int maxRows) {
+        return limit(maxRows);
     }
 
     /**
      * Configures the loader to select a sample of the rows from the ResultSet. Unlike
      * {@link DataFrame#sampleRows(int, Random)}, this method can be used on potentially very large
-     * result sets. If you are executing multiple sampling runs in parallel, consider using {@link #sampleRows(int, Random)},
+     * result sets. If you are executing multiple sampling runs in parallel, consider using {@link #rowsSample(int, Random)},
      * as this method is using a shared {@link Random} instance with synchronization.
      *
      * @param size the size of the sample. Can be bigger than the result set size (as the result set size is not known upfront).
      * @return this loader instance
-     * @since 0.7
+     * @since 1.0.0-M20
      */
+    public SqlLoader rowsSample(int size) {
+        return rowsSample(size, Sampler.getDefaultRandom());
+    }
+
+    /**
+     * @deprecated in facfor of {@link #rowsSample(int)}
+     */
+    @Deprecated(since = "1.0.0-M20", forRemoval = true)
     public SqlLoader sampleRows(int size) {
-        return sampleRows(size, Sampler.getDefaultRandom());
+        return rowsSample(size);
     }
 
     /**
@@ -74,9 +93,9 @@ public class SqlLoader {
      * @param size   the size of the sample. Can be bigger than the result set size (as the result set size is not known upfront).
      * @param random a custom random number generator
      * @return this loader instance
-     * @since 0.7
+     * @since 1.0.0-M20
      */
-    public SqlLoader sampleRows(int size, Random random) {
+    public SqlLoader rowsSample(int size, Random random) {
 
         if (this.rowSampleSize == size && this.rowsSampleRandom == random) {
             return this;
@@ -86,6 +105,14 @@ public class SqlLoader {
         copy.rowSampleSize = size;
         copy.rowsSampleRandom = random;
         return copy;
+    }
+
+    /**
+     * @deprecated in favor of {@link #rowsSample(int, Random)}
+     */
+    @Deprecated(since = "1.0.0-M20", forRemoval = true)
+    public SqlLoader sampleRows(int size, Random random) {
+        return rowsSample(size, random);
     }
 
     /**
@@ -123,7 +150,7 @@ public class SqlLoader {
                 .sampleRows(rowSampleSize, rowsSampleRandom)
                 .appender();
 
-        return new SqlLoaderWorker(appender, maxRows).load(rs);
+        return new SqlLoaderWorker(appender, limit).load(rs);
     }
 
     protected Index createIndex(ResultSet rs) throws SQLException {